fix integration LDAP tests and split into individual files
