+++ it/com/yetu/oauth2provider/services/LDAPPermissionsService.scala
@@ -1,0 +1,21 @@
+package com.yetu.oauth2provider.services
+
+import com.yetu.oauth2provider.oauth2.models.ClientPermission
+import securesocial.core.services.SaveMode
+
+//TODO: implement permissions correctly and make sure this test leaves no traces behind.
+class LDAPPermissionsService extends LDAPBaseSpec {
+//  "The LDAP permission service" must {
+//    val clientPermission = ClientPermission("123456", Some(List("scope1")))
+//    "delete, store and retrieve a permissions " in {
+//      personService.deleteUser(testUser.userId)
+//      permissionService.deletePermission(testUser.userId, clientPermission.clientId)
+//      personService.save(testUser.toBasicProfile, SaveMode.SignUp)
+//      permissionService.savePermission(testUser.userId, clientPermission, true)
+//      val retrieved = permissionService.findPermission(testUser.userId, clientPermission.clientId)
+//      retrieved.get mustEqual clientPermission
+//      permissionService.deletePermission(testUser.userId, clientPermission.clientId)
+//      personService.deleteUser(testUser.userId)
+//    }
+//  }
+}
